correctd value provider

+++ src/test/java/org/ssu/edu/teachua/utils/providers/DataProviderChallenge.java
@@ -98,7 +98,7 @@ public class DataProviderChallenge {
     @DataProvider(name = "dpTestIfChallengeIsNotCreated")
     public static Object[][] dpTestIfChallengeNotCreated() {
         return new Object[][]{
-                {"testTitle", "Lorem Ipsum is simply dummy text of the printing and typesetting industry.", null, "/upload/photos/image.png", 546789,
+                {"testTitle", "Lorem Ipsum is simply dummy text of the printing and typesetting industry.", null, "/upload/photos/image.png", BigInteger.valueOf(546789),
                         Arrays.asList("Дыผð*.:", "t", "Lorem Ipsum is simply dummy text of the printing and typesetting industry.", ""),
                         Arrays.asList("Це поле може містити тільки українські та англійські літери, цифри та спеціальні символи",
                                 "Назва Челенджу закоротка",
